fix: restructure mteb tasks  to its own folder
