+++ src/seb/mteb_tasks/__init__.py
@@ -1,0 +1,4 @@
+from .classification import *
+from .clustering import *
+from .retrieval import *
+from .bitext_mining import *
