+++ src/seb/mteb_tasks/bitext_mining/__init__.py
@@ -1,0 +1,1 @@
+from .norwegian_courts_bitext_mining import NorwegianCourtsBitextMining
