+++ src/seb/mteb_tasks/classification/__init__.py
@@ -1,0 +1,1 @@
+from .norwegian_parliament_classification import NorwegianParliamentClassification
