+++ src/seb/mteb_tasks/clustering/__init__.py
@@ -1,0 +1,2 @@
+from .swedn_clustering import SwednClustering
+from .vg_clustering import VGClustering
