+++ src/seb/mteb_tasks/retrieval/__init__.py
@@ -1,0 +1,4 @@
+from .dan_fever import DanFever
+from .swefaq_retrieval import SweFaqRetrieval
+from .swedn_retrieval import SwednRetrieval
+from .swefaq_retrieval import SweFaqRetrieval
