+++ src/seb/registered_tasks/multilingual.py
@@ -1,6 +1,5 @@
 from datetime import datetime
 from functools import partial
-from typing import Any
 
 import numpy as np
 from datasets import DatasetDict, concatenate_datasets
